+++ codec/codec.go
@@ -16,6 +16,7 @@ const (
 var (
 	ErrHeaderNotEnough = errors.New("header buffer too short")
 	ErrBodyNotEnough   = errors.New("body buffer too short")
+	ErrJavaException   = errors.New("got java exception")
 	ErrIllegalPackage  = errors.New("illegal package!")
 )